src/microservices/astack-authentication/api/main/requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
ecdsa==0.17.0
fastapi==0.78.0
jose==1.0.0
loguru==0.6.0
mock==4.0.3
passlib==1.7.4
pg8000==1.16.6
python-multipart==0.0.5
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-jose==3.3.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing==3.7.4.3
typing_extensions==4.2.0
uuid==1.30
uvicorn==0.17.6
redis==4.3.4